Revert "feat(module-index-server): add schema data to module index table before backfill"
